rename AnalysisSettings.py to settings.py
