+++ biahub/segment.py
@@ -8,7 +8,6 @@ import torch
 from iohub.ngff import open_ome_zarr
 from iohub.ngff.utils import create_empty_plate, process_single_position
 
-from biahub.AnalysisSettings import SegmentationSettings
 from biahub.cli import utils
 from biahub.cli.monitor import monitor_jobs
 from biahub.cli.parsing import (
@@ -20,6 +19,7 @@ from biahub.cli.parsing import (
     sbatch_to_submitit,
 )
 from biahub.cli.utils import estimate_resources, yaml_to_model
+from biahub.settings import SegmentationSettings
 
 
 def segment_data(
